seperate note card from edit page
