+++ frontend/src/features/playScenario/modals/NotesModal/Note.module.scss
